move APIFnGetLinks to action_getlinks.go
